refactor(web): send components/ui's single-feature files home (LFE-14804)
Three files in the frozen components/ui folder are each used by exactly one
feature, so they belong in it:

  object-not-found-card.tsx -> features/annotation-queues/components
  timeline.tsx              -> features/prompts/components
  dialog-controller.tsx     -> features/in-app-agent/components

Done with `pnpm structure:move`: git mv plus the importer rewrite, so the moved
files are byte-identical and each has exactly one import line updated.

rule 6 (a file used by one feature lives in it): 103 -> 100
rule 13 (components/ui is frozen):               118 -> 115
every other rule unchanged; total 2692 -> 2686.
