+++ web/src/features/annotation-queues/components/AnnotationQueueItemPage.tsx
@@ -37,7 +37,7 @@ import { useAnnotationQueueData } from "./shared/hooks/useAnnotationQueueData";
 import { useAnnotationObjectData } from "./shared/hooks/useAnnotationObjectData";
 import { TraceAnnotationProcessor } from "./processors/TraceAnnotationProcessor";
 import { SessionAnnotationProcessor } from "./processors/SessionAnnotationProcessor";
-import { ObjectNotFoundCard } from "@/src/components/ui/object-not-found-card";
+import { ObjectNotFoundCard } from "@/src/features/annotation-queues/components/object-not-found-card";
 import { useSession } from "next-auth/react";
 
 // A single row in the keyboard-shortcuts cheatsheet: label on the left, one or
