+++ web/src/features/in-app-agent/components/InAppAgentWindowHost.tsx
@@ -3,7 +3,7 @@
 import { useLayoutEffect, useRef } from "react";
 
 import { ConfirmDialog } from "@/src/components/ui/confirm-dialog";
-import { DialogController } from "@/src/components/ui/dialog-controller";
+import { DialogController } from "@/src/features/in-app-agent/components/dialog-controller";
 import { Layer } from "@/src/components/ui/layer";
 import { ControlledInAppAgentWindow } from "@/src/features/in-app-agent/components/ControlledInAppAgentWindow";
 import type { InAppAgentWindowConversation } from "@/src/features/in-app-agent/components/InAppAgentWindow";
